Refactors domain alias tests.
